+++ src/components/pages/Stake/reducer.ts
@@ -233,9 +233,11 @@ const calculate = (state: State): State => {
       slope.mul(10000).mul(Math.floor(Math.sqrt(length))),
     );
 
-    const simulatedTotalStaticWeight = incentivisedVotingLockup.totalStaticWeight.add(
-      simulatedStakingBalance,
-    );
+    const simulatedTotalStaticWeight = incentivisedVotingLockup.totalStaticWeight
+      .add(simulatedStakingBalance)
+      // With an existing balance, remove the existing staking balance,
+      // because it's already included in the simulated balance.
+      .sub(userStakingBalance ?? new BigDecimal(0));
 
     const simulatedApy = getShareAndAPY(
       rewardRate,